+++ DbzGame.py
@@ -1,4 +1,7 @@
-from arguments import pychoose
-from Functions import quest, scoreTotal
+from arguments import pydb, pydbz, pydbs
+from Functions import scoreTotal, Game
+import random
 
+Game(pydb, pydbz, pydbs)
 
+scoreTotal()