+++ arguments.py
@@ -16,7 +16,7 @@ piccolo = zFighter("piccolo", "Db", True)
 roshi = zFighter("Master Roshi", "Db", True)
 
 #Newly imoported Dbz Characters
-nappa = zFighter("nappa", "Dbz", True)
+nappa = zFighter("Nappa", "Dbz", True)
 futureTrunks = zFighter("Future Trunks","Dbz", True)
 raddiz = zFighter("Raddiz", "Dbz", True)
 cell = zFighter("Cell","Dbz", True)
@@ -25,26 +25,32 @@ frieza = zFighter("Frieza","Dbz", True)
 broly = zFighter("Broly", "Dbz", True) 
 
 #Newly imported Dbs Characters
-bulla = zFighter("bulla", "Dbs", False)
+bulla = zFighter("Bulla", "Dbs", False)
 beerus = zFighter("Beerus", "Dbs", True)
 frost = zFighter("Frost", "Dbs", True)
 hit = zFighter("Hit","Dbs",True)
 jiren = zFighter("Jiren","Dbs",True)
-zamatsu = zFighter("Zamtsu","Dbs",True)
+zamatsu = zFighter("Zamatsu","Dbs",True)
 zeno = zFighter("Lord Zeno","Dbs",False)
 vados = zFighter("Vados","Dbs",True)
-choompa = zFighter("Choompa","Dbs",True)
-
-
-
-
+chompa = zFighter("Chompa","Dbs",True)
 
 #These are the player's basic arguments; user attempts, score, guess limit
 playerargs = parameter(0,0,4)
 
-#This is a list of the created charcter breakdowns
+#This is a list of the created charcter breakdowns of Dragonball
+pydb = random.choice([goku, krillin, bulma, 
+korrin, chichi, oolong, 
+A8, piccolo, roshi,])
 
-#This is a list of newly created characters
+#This is a list of the created charcter breakdowns of Dragonball z
+pydbz = random.choice([vegeta, nappa, futureTrunks, 
+raddiz, cell, vegito, 
+frieza, broly,])
 
+#This is a list of the created charcter breakdowns of Dragonball Super
+pydbs = random.choice([bulla, beerus, frost, 
+hit, jiren, zamatsu, 
+zeno, vados, chompa,])
 
-#This is the random module selecting characters from the above list
+#This is the random module selecting characters from the above lists