+++ changelog.md
@@ -1,12 +1,22 @@
+v 6.85
+Functions
+-------------------
+- imported random
+- created new function called Game
+- deleted break from quest function incorrect answer block to allow for retry
+
+
+
 v 6.75
 arguments module
 -------------------
-- added new character arguments (cell, frost, F. Trunks)
+- added new character arguments (cell, frost, F. Trunks, raddiz, nappa, vegito, frieza, broly, frost, hit, jiren, zamatsu, zeno, vados, chompa)
 - imported random
-- created list
+- created lists for Db, Dbz, Dbs characters
 - added to the list with newly created characters
 - added more #comments/information zones
 - organized characters in sections by show
+- implement random.choice to list
 
 
 classes module